v.1.1.9.6
Failing the boogeyman quest on Limited Life now drops your time properly.

World border is now only set to 500x500 if it has not been set to anything before.

Added custom recipes for saddles and name tags.

+++ src/main/java/net/mat0u5/lifeseries/series/BoogeymanManager.java
@@ -170,12 +170,17 @@ public class BoogeymanManager {
                             "That means their lives have not been set to 1. You must do this manually once they are online again."));
                     continue;
                 }
-                player.sendMessage(Text.of("§7You failed to kill a green or yellow name last session as the Boogeyman. As punishment, you have dropped to your §cLast Life§7. " +
-                        "All alliances are severed and you are now hostile to all players. You may team with others on their Last Life if you wish."));
-                OtherUtils.broadcastMessage(player.getStyledDisplayName().copy().append(Text.of("§7 failed to kill a player while being the §cBoogeyman§7. They have been been dropped to their §cLast Life§7")));
-                currentSeries.setPlayerLives(player, 1);
+                playerFailBoogeyman(player);
             }
         }
+    }
+    public void playerFailBoogeyman(ServerPlayerEntity player) {
+        if (!currentSeries.isAlive(player)) return;
+        if (currentSeries.isOnLastLife(player)) return;
+        player.sendMessage(Text.of("§7You failed to kill a green or yellow name last session as the Boogeyman. As punishment, you have dropped to your §cLast Life§7. " +
+                "All alliances are severed and you are now hostile to all players. You may team with others on their Last Life if you wish."));
+        OtherUtils.broadcastMessage(player.getStyledDisplayName().copy().append(Text.of("§7 failed to kill a player while being the §cBoogeyman§7. They have been dropped to their §cLast Life§7")));
+        currentSeries.setPlayerLives(player, 1);
     }
     public void playerLostAllLives(ServerPlayerEntity player) {
         Boogeyman boogeyman = getBoogeyman(player);
